instagram analytics/requirements.txt
# Core Python Libraries
facebook-sdk==3.1.0
requests==2.31.0
python-dotenv==1.0.0

# Data Processing
pandas==2.1.4
matplotlib==3.8.2

# Visualization & Formatting
tabulate==0.9.0

# Recommended development setup
pylint==3.0.3
black==24.1.1

# Recommended Python version: 3.8+